ET_Mascota_Avila_Coñuenao_003D

+++ app/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import redirect, render
-from django.contrib.auth.models import User
-from app.models import Cliente, Especie, Mascota
+from django.contrib.auth.models import User 
+from app.models import Cliente, Especie, Mascota, CarritoCliente
 
 
 # Create your views here.
@@ -17,15 +17,34 @@ def galeria(request):
     else:
         return render(request,'galeria.html')
        
+def carrito(request):
+    if(not request.user.is_authenticated):
+        return redirect('login')
+    else:
+            if (request.user.is_superuser):
+                user = User.objects.all()
+                return render(request,"carrito.html",{"clientes":user})
+
+            else:
+                carrito=CarritoCliente.objects.filter(id_user=request.user.id)
+                datos={
+                    "carrito":list(map(lambda x: x.id_mascota, carrito)),
+                    "carro":list(map(lambda x: x, carrito))
+                }
+                
+                
+                if (len( datos["carrito"] ) != 0):
+                    return render(request,"carrito.html",datos)
+                else:
+                    return render(request,"carrito.html")
+
 
 def equipo(request):
     return render(request,"equipo.html")
 
 def register(request):
-    if(not request.user.is_authenticated):
-        return redirect('login')
-    else:
-        return render(request,"register.html")
+    
+    return render(request,"register.html")
 
 def login(request):
     return render(request,"login.html")
@@ -38,26 +57,32 @@ def mascotas(request):
     if(not request.user.is_authenticated):
        return redirect('login')
     else:
-        return render(request, "mascota.html", {"especie": especie})
+        if(request.user.is_superuser == True):
+            return render(request, "mascota.html", {"especie": especie})
+        else:
+            return redirect('index')
 
 def añadirMascota(request):
     if(not request.user.is_authenticated):
         return redirect('login')
     else:
-        if request.method == 'POST':
+        if(request.user.is_superuser == True):
+            if request.method == 'POST':
 
-            mascota= Mascota()
-            mascota.nombreMascota=request.POST.get("nombre_Mascota")
-            mascota.nombreEspecie=Especie.objects.get(nombreEspecie=request.POST.get("nombre_Especie") )
-            mascota.raza=request.POST.get("raza_Mascota")
-            mascota.precio=request.POST.get("precio_mascota")
-            mascota.nombreDueño=request.POST.get("nombre_Dueño")
+                mascota= Mascota()
+                mascota.nombreMascota=request.POST.get("nombre_Mascota")
+                mascota.nombreEspecie=Especie.objects.get(nombreEspecie=request.POST.get("nombre_Especie") )
+                mascota.raza=request.POST.get("raza_Mascota")
+                mascota.precio=request.POST.get("precio_mascota")
+                mascota.nombreDueño=request.POST.get("nombre_Dueño")
 
-            if len(request.FILES)!=0:
-                mascota.imagenMascota=request.FILES["imagen_mascota"]
+                if len(request.FILES)!=0:
+                    mascota.imagenMascota=request.FILES["imagen_mascota"]
 
-            mascota.save()
-            return redirect('galeria')
+                mascota.save()
+                return redirect('galeria')
+        else:
+            return redirect('index')
             
 
 def eliminarMascota(request,id):
@@ -65,9 +90,12 @@ def eliminarMascota(request,id):
         return redirect('login')
         
     else:
-        mascota= Mascota.objects.get(id_mascota=id)
-        mascota.delete()
-        return redirect('galeria')
+        if(request.user.is_superuser == True):
+            mascota= Mascota.objects.get(id_mascota=id)
+            mascota.delete()
+            return redirect('galeria')
+        else:
+            return redirect('index')
 
 
 def editarMascota(request,id):
@@ -82,37 +110,41 @@ def editarMascota(request,id):
         return redirect('login')
 
     else:
-        if request.method == 'POST':
-            mascota= Mascota.objects.get(id_mascota=id)
-            mascota.nombreMascota=request.POST.get("nombre_Mascota")
+        if(request.user.is_superuser == True):
+            if request.method == 'POST':
+                mascota= Mascota.objects.get(id_mascota=id)
+                mascota.nombreMascota=request.POST.get("nombre_Mascota")
 
-            if(mascota.nombreEspecie == mascota.nombreEspecie):
-                mascota.nombreEspecie= Especie.objects.get(nombreEspecie=request.POST.get("nombre_Especie") )
-            else:
-                pass
-            mascota.raza=request.POST.get("raza_Mascota")
-            mascota.precio=request.POST.get("precio_mascota")
-            mascota.nombreDueño=request.POST.get("nombre_Dueño")
+                if(mascota.nombreEspecie == mascota.nombreEspecie):
+                    mascota.nombreEspecie= Especie.objects.get(nombreEspecie=request.POST.get("nombre_Especie") )
+                else:
+                    pass
+                mascota.raza=request.POST.get("raza_Mascota")
+                mascota.precio=request.POST.get("precio_mascota")
+                mascota.nombreDueño=request.POST.get("nombre_Dueño")
 
-            if len(request.FILES)!=0:
-                mascota.imagenMascota=request.FILES["imagen_mascota"]
-                
-            mascota.save()
-            return redirect('galeria')
-        
-        return render(request,"editMascota.html",data)
+                if len(request.FILES)!=0:
+                    mascota.imagenMascota=request.FILES["imagen_mascota"]
+                    
+                mascota.save()
+                return redirect('galeria')
+            
+            return render(request,"editMascota.html",data)
+        else:
+            return redirect('index')
    
 def cliente(request):
     if(not request.user.is_authenticated):
         return redirect('login')
     else:
-        cliente=Cliente.objects.all()
-        return render(request,"cliente.html",{"cliente":cliente})
+        if(request.user.is_superuser == True):
+            cliente=Cliente.objects.all()
+            return render(request,"cliente.html",{"cliente":cliente})
+        else:
+            return redirect('index')
 
 def añadirCliente(request):
-    if(not request.user.is_authenticated):
-        return redirect('login')
-    else:
+
         if request.method == 'POST':
 
             cliente= Cliente()
@@ -121,19 +153,29 @@ def añadirCliente(request):
             cliente.direccion=request.POST.get("direccion_cliente")
             cliente.telefono=request.POST.get("telefono_cliente")
             cliente.password=request.POST.get("contraseña_cliente")
-        
             cliente.save()
+
+            if cliente.password == request.POST.get("contraseña_cliente2"):
+                if User.objects.filter(email=cliente.email).exists():
+                    return render(request,"register.html",{"error":"El correo ya existe"})
+                else:
+                    user = User.objects.create_user(username=cliente.nombre, password=cliente.password,email=cliente.email)
+                    user.save();
+
             return redirect('cliente')
             
 
 def eliminarCliente(request,id):
+    
     if(not request.user.is_authenticated):
         return redirect('login')
-        
     else:
-        cliente= Cliente.objects.get(id_cliente=id)
-        cliente.delete()
-        return redirect('cliente')
+        if(request.user.is_superuser == True):
+            cliente= Cliente.objects.get(id_cliente=id)
+            cliente.delete()
+            return redirect('cliente')
+        else:
+            return redirect('index')
 
 
 def editarCliente(request,id):
@@ -141,18 +183,114 @@ def editarCliente(request,id):
     if(not request.user.is_authenticated):
         return redirect('login')
     else:
-        if request.method == 'POST':
-            
-            cliente= Cliente.objects.get(id_cliente=id)
-            cliente.nombre=request.POST.get("nombre_cliente")
-            cliente.direccion=request.POST.get("direccion_cliente")
-            cliente.telefono=request.POST.get("telefono_cliente")
-            cliente.email=request.POST.get("email_cliente")
-            cliente.password=request.POST.get("contraseña_cliente")
-            
-            cliente.save()
+        if(request.user.is_superuser == True):
+            if request.method == 'POST':
+                
+                cliente= Cliente.objects.get(id_cliente=id)
+                cliente.nombre=request.POST.get("nombre_cliente")
+                cliente.direccion=request.POST.get("direccion_cliente")
+                cliente.telefono=request.POST.get("telefono_cliente")
+                cliente.email=request.POST.get("email_cliente")
+                cliente.password=request.POST.get("contraseña_cliente")
+                
+                cliente.save()
 
-            return redirect('cliente')
-        return render(request,"editCliente.html",{"cliente":cliente})    
+                return redirect('cliente')
+            return render(request,"editCliente.html",{"cliente":cliente})    
+        else:
+            return redirect('index')
+
+
+
+def carro(request,id):
+    if(not request.user.is_authenticated):
+        return redirect('login')
+    else:
+        mascota = Mascota.objects.get(id_mascota=id)
+        carrito=CarritoCliente.objects.filter(id_user=request.user.id)
+        carro=list(map(lambda x: x.id_mascota, carrito))
+
+
+        if mascota in carro:
+            print("ya esta en el carrito")
+            return redirect('galeria') 
+        else:
+            carrito=CarritoCliente()
+            carrito.id_user=User.objects.get(id=request.user.id)
+            carrito.id_mascota=mascota
+            carrito.save()
+            
+            print("agregado")
+            return redirect('galeria') 
+
+def carro_eliminar(request,id):
+    if(not request.user.is_authenticated):
+        return redirect('login')
+    else:
+        if(request.user.is_superuser == True):
+            carrito=CarritoCliente.objects.get(id_user=request.user.id,id_mascota=id)
+            carrito.delete()
+            
+
+            return redirect('carrito')
+        else:
+            return redirect('index')
+
+
         
+def lista_carrito(request,id):
+    if(not request.user.is_authenticated):
+        return redirect('login')
+    else:
+        if(request.user.is_superuser == True):
+            carrito=CarritoCliente.objects.filter(id_user=id)
+            
+            
+            datos={
+                "carrito":list(map(lambda x: x.id_mascota, carrito)),
+                "carro":list(map(lambda x: x, carrito))
+            }
+            
 
+            if (len( datos["carrito"] ) != 0):
+                    return render(request,"listaCarrito.html", datos)
+            else:
+                return render(request,"carrito.html")
+        else:
+            return redirect('index')
+    
+def seguimiento(request,id,slug):
+    if(not request.user.is_authenticated):
+        return redirect('login')
+    else:
+        if(request.user.is_superuser == True):
+
+            carrito=CarritoCliente.objects.filter(id_user=id,id_mascota=slug)
+           
+            if request.POST.get("select_seguimiento") == "En camino":
+                carrito.update(seguimiento="En camino")
+                return redirect('carrito')
+
+            elif request.POST.get("select_seguimiento") == "Entregado":
+                carrito.update(seguimiento="Entregado")
+                return redirect('carrito')
+
+            elif request.POST.get("select_seguimiento") == "Preparacion":
+                carrito.update(seguimiento="Preparacion")
+                return redirect('carrito')
+        
+        
+           
+        else:
+            return redirect('index')
+
+def elimarcarrouser(request,id,slug):
+    if(not request.user.is_authenticated):
+        return redirect('login')
+    else:
+        if(request.user.is_superuser == True):
+            carrito=CarritoCliente.objects.get(id_user=id,id_mascota=slug)
+            carrito.delete()
+            return redirect('carrito')
+        else:
+            return redirect('index')
